Add a rebuild switch and bash completion for "repo"

+++ run.sh
@@ -2,8 +2,8 @@
 
 cd $(dirname $0)
 
-SOURCE=$(pwd)/android
-CCACHE=$(pwd)/ccache
+SOURCE=$(readlink -f $(pwd)/android)
+CCACHE=$(readlink -f $(pwd)/ccache)
 CONTAINER_HOME=/home/cmbuild
 CONTAINER=cyanogenmod
 REPOSITORY=stucki/cyanogenmod
@@ -27,6 +27,23 @@ while [[ $# > 0 ]]; do
 	shift
 done
 
+
+while [[ $# > 0 ]]
+do
+key="$1"
+
+case $key in
+    -r|--rebuild)
+    FORCE_BUILD=1
+    ;;
+    *)
+    shift # past argument or value
+    ;;
+esac
+shift
+
+done
+
 # Create shared folders
 mkdir -p $SOURCE
 mkdir -p $CCACHE
@@ -43,12 +60,17 @@ elif [[ $FORCE_BUILD = 1 ]] || ! echo "$IMAGE_EXISTS" | grep -q "$TAG"; then
 
 	echo "Building Docker image $REPOSITORY:$TAG..."
 	docker build -t $REPOSITORY:$TAG .
+	OK=$?
 
 	# After successful build, delete existing containers
 	IS_EXISTING=$(docker inspect -f '{{.Id}}' $CONTAINER 2>/dev/null)
-	if [[ -n "$IS_EXISTING" ]]; then
+	if [[ $OK -eq 0 ]] && [[ -n "$IS_EXISTING" ]]; then
 		docker rm $CONTAINER
 	fi
+fi
+
+if [[ $OK -ne 0 ]]; then
+	exit 1;
 fi
 
 # With the given name $CONTAINER, reconnect to running container, start
